Separate driver rustc and librustc
rustc is now a minimal wrapper of librustc.
