+++ Makefile.in
@@ -209,8 +209,11 @@ STDLIB_INPUTS := $(wildcard $(addprefix $(S)src/libstd/,          \
 ######################################################################
 
 COMPILER_CRATE := $(S)src/comp/rustc.rc
-COMPILER_INPUTS := $(wildcard $(addprefix $(S)src/comp/, \
-                                rustc.rc *.rs */*.rs */*/*.rs))
+COMPILER_INPUTS := $(filter-out $(S)src/comp/driver/rustc.rs,         \
+				$(wildcard $(addprefix $(S)src/comp/, \
+                                rustc.rc *.rs */*.rs */*/*.rs)))
+
+RUSTC_INPUTS := $(S)src/comp/driver/rustc.rs
 
 ######################################################################
 # LLVM macros
