+++ mk/target.mk
@@ -46,17 +46,14 @@ $$(TLIB$(1)_T_$(2)_H_$(3))/$$(CFG_RUSTLLVM): \
 	$$(Q)cp $$< $$@
 
 $$(TBIN$(1)_T_$(2)_H_$(3))/rustc$$(X):				\
-		$$(COMPILER_CRATE) $$(COMPILER_INPUTS)		\
-		$$(TSREQ$(1)_T_$(2)_H_$(3))					\
-		$$(TLIB$(1)_T_$(2)_H_$(3))/$$(CFG_RUSTLLVM)	\
-		$$(TCORELIB_DEFAULT$(1)_T_$(2)_H_$(3))      \
-		$$(TSTDLIB_DEFAULT$(1)_T_$(2)_H_$(3))
+		$$(RUSTC_INPUTS)                                \
+		$$(TLIBRUSTC_DEFAULT$(1)_T_$(2)_H_$(3))
 	@$$(call E, compile_and_link: $$@)
 	$$(STAGE$(1)_T_$(2)_H_$(3))  -o $$@ $$<
 
 $$(TLIB$(1)_T_$(2)_H_$(3))/$$(CFG_LIBRUSTC):		\
 		$$(COMPILER_CRATE) $$(COMPILER_INPUTS)		\
-		$$(TSREQ$(1)_T_$(2)_H_$(3))					\
+		$$(TSREQ$(1)_T_$(2)_H_$(3))			\
 		$$(TLIB$(1)_T_$(2)_H_$(3))/$$(CFG_RUSTLLVM)	\
 		$$(TCORELIB_DEFAULT$(1)_T_$(2)_H_$(3))      \
 		$$(TSTDLIB_DEFAULT$(1)_T_$(2)_H_$(3))
